feat(widgets): handle logged out and errors for TopicRecommendationsWidget
