+++ ItemWidgets/ItemWidgetsKit/Models/ItemsListEntry.swift
@@ -2,20 +2,46 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+import SharedPocketKit
 import Textile
 import WidgetKit
 
 struct ItemsListEntry: TimelineEntry {
     let date: Date
-    let name: String
     let titleColor: ColorAsset
-    let contentType: ItemsListContentType
+    let content: ItemsWidgetContent
 
     init(date: Date, name: String, titleColor: ColorAsset = .ui.coral2, contentType: ItemsListContentType) {
         self.date = date
-        self.name = name
         self.titleColor = titleColor
-        self.contentType = contentType
+        self.content = ItemsWidgetContent(name: name, contentType: contentType)
+    }
+}
+
+struct ItemsWidgetContent {
+    let name: String
+    let contentType: ItemsListContentType
+}
+
+extension ItemsWidgetContent {
+    static var sampleContent: ItemsWidgetContent {
+        ItemsWidgetContent(
+            name: "Sample Topic",
+            contentType: .items(
+                [
+                    ItemRowContent(
+                        content: ItemContent(
+                            url: "https://getPocket.com",
+                            title: "Sample Entry",
+                            imageUrl: nil,
+                            bestDomain: "https://getPocket.com",
+                            timeToRead: 0
+                        ),
+                        image: nil
+                    )
+                ]
+            )
+        )
     }
 }
